requirements.txt
streamlit>=1.35.0
streamlit-mic-recorder>=0.0.1
groq>=0.4.0
python-dotenv>=1.0.0
tabulate>=0.9.0
gtts>=2.4.0
